+++ src/frontend/token.zig
@@ -165,6 +165,7 @@ pub const Tag = enum {
     kw_match,
     kw_return,
     kw_exit,
+    kw_yield,
 
     // Module/interop keywords
     /// Module import keyword.
@@ -252,6 +253,7 @@ pub const Tag = enum {
             .kw_match => "match",
             .kw_return => "return",
             .kw_exit => "exit",
+            .kw_yield => "yield",
             .kw_import => "import",
             .kw_try => "try",
             .kw_catch => "catch",
@@ -294,6 +296,7 @@ const keyword_map = std.StaticStringMap(Tag).initComptime(.{
     .{ "match", .kw_match },
     .{ "return", .kw_return },
     .{ "exit", .kw_exit },
+    .{ "yield", .kw_yield },
     .{ "import", .kw_import },
     .{ "try", .kw_try },
     .{ "catch", .kw_catch },